Move setParent logic to useNote
