+++ src/application/services/useNote.ts
@@ -71,6 +71,13 @@ interface UseNoteComposableState {
   unlinkParent: () => Promise<void>;
 
   /**
+   * Set parent for the note
+   * @param id - Child note id
+   * @param newParentURL - New parent note URL
+   */
+  setParent: (id: NoteId, newParentURL: string) => Promise<void>;
+
+  /**
    * Defines if user can edit note
    */
   canEdit: Ref<boolean>;
@@ -271,6 +278,22 @@ export default function (options: UseNoteComposableOptions): UseNoteComposableSt
   }
 
   /**
+   * Set parent for the note
+   * @param id - Child note id
+   * @param newParentURL - New parent note URL
+   */
+  const setParent = async (id: NoteId, newParentURL: string): Promise<void> => {
+    try {
+      parentNote.value = await noteService.setParentByUrl(id, newParentURL);
+    } catch (error) {
+      if (error instanceof Error) {
+        window.alert(error.message);
+      }
+      throw error;
+    }
+  };
+
+  /**
    * Get note by custom hostname
    */
   const resolveHostname = async (): Promise<void> => {
@@ -327,6 +350,7 @@ export default function (options: UseNoteComposableOptions): UseNoteComposableSt
     resolveToolsByContent,
     save,
     unlinkParent,
+    setParent,
     parentNote,
   };
 }
